+++ cmd/rqlited/main.go
@@ -62,6 +62,7 @@ var discoID string
 var expvar bool
 var pprofEnabled bool
 var onDisk bool
+var fkConstraints bool
 var raftLogLevel string
 var raftNonVoter bool
 var raftSnapThreshold uint64
@@ -109,6 +110,7 @@ func init() {
 	flag.BoolVar(&expvar, "expvar", true, "Serve expvar data on HTTP server")
 	flag.BoolVar(&pprofEnabled, "pprof", true, "Serve pprof data on HTTP server")
 	flag.BoolVar(&onDisk, "on-disk", false, "Use an on-disk SQLite database")
+	flag.BoolVar(&fkConstraints, "fk", false, "Enable SQLite foreign key constraints")
 	flag.BoolVar(&showVersion, "version", false, "Show version information and exit")
 	flag.BoolVar(&raftNonVoter, "raft-non-voter", false, "Configure as non-voting node")
 	flag.StringVar(&raftHeartbeatTimeout, "raft-timeout", "1s", "Raft heartbeat timeout")
@@ -196,6 +198,7 @@ func main() {
 		log.Fatalf("failed to determine absolute data path: %s", err.Error())
 	}
 	dbConf := store.NewDBConfig(!onDisk)
+	dbConf.FKConstraints = fkConstraints
 
 	str := store.New(raftTn, &store.StoreConfig{
 		DBConf: dbConf,