+++ store/db_config.go
@@ -2,7 +2,11 @@ package store
 
 // DBConfig represents the configuration of the underlying SQLite database.
 type DBConfig struct {
-	Memory bool // Whether the database is in-memory only.
+	// Whether the database is in-memory only.
+	Memory bool `json:"memory"`
+
+	// Enforce Foreign Key constraints
+	FKConstraints bool `json:"fk_constraints"`
 }
 
 // NewDBConfig returns a new DB config instance.